admin/src/requirements.txt
streamlit==1.34.0
redis==4.6.0
python-dotenv==1.0.1